third_party/blink/web_tests/fast/dom/Window/anonymous-slot-with-changes-expected.txt

Tests that we clone object hierarchies
PASS: 'postMessage((function(){}))' threw DataCloneError: Failed to execute 'postMessage' on 'Window': function(){} could not be cloned.
PASS: eventData is null of type object
PASS: eventData is null of type object
PASS: eventData is null of type object
PASS: eventData is null of type object
PASS: eventData is null of type object
PASS: eventData is undefined of type undefined
PASS: eventData is undefined of type undefined
PASS: eventData is undefined of type undefined
PASS: eventData is undefined of type undefined
PASS: eventData is undefined of type undefined
PASS: eventData is 1 of type number
PASS: eventData is 1 of type number
PASS: eventData is 1 of type number
PASS: eventData is 1 of type number
PASS: eventData is 1 of type number
PASS: eventData is true of type boolean
PASS: eventData is true of type boolean
PASS: eventData is true of type boolean
PASS: eventData is true of type boolean
PASS: eventData is true of type boolean
PASS: eventData is 1 of type string
PASS: eventData is 1 of type string
PASS: eventData is 1 of type string
PASS: eventData is 1 of type string
PASS: eventData is 1 of type string
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is of type object
PASS: eventData is of type object
PASS: eventData is of type object
PASS: eventData is of type object
PASS: eventData is of type object
PASS: eventData is 1,2,3 of type object
PASS: eventData is 1,2,3 of type object
PASS: eventData is 1,2,3 of type object
PASS: eventData is 1,2,3 of type object
PASS: eventData is 1,2,3 of type object
PASS: eventData is ,,1 of type object
PASS: eventData is ,,1 of type object
PASS: eventData is ,,1 of type object
PASS: eventData is ,,1 of type object
PASS: eventData is ,,1 of type object
PASS: eventData is 2009-02-13T23:31:30.000Z of type object
PASS: eventData is 2009-02-13T23:31:30.000Z of type object
PASS: eventData is 2009-02-13T23:31:30.000Z of type object
PASS: eventData is 2009-02-13T23:31:30.000Z of type object
PASS: eventData is 2009-02-13T23:31:30.000Z of type object
PASS: eventData is done of type string
PASS: eventData is done of type string
PASS: eventData is done of type string
PASS: eventData is done of type string
PASS: eventData is done of type string
